% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{act_data_df}
\alias{act_data_df}
\title{Get account balances, positions, or orders returned as a data frame}
\usage{
act_data_df(
  dataType = c("balances", "positions", "orders"),
  accessToken = NULL
)
}
\arguments{
\item{dataType}{'balances' for current cash balances, 'positions' for current
account positions, 'orders' for orders entered on the current day. Default
is set to 'balances'.}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
a data frame of requested account details
}
\description{
Retrieves a data frame of data for the accounts linked to the Access Token
}
\details{
The output will be a data frame of requested details for TD Ameritrade
accounts linked to the access token. Use \code{\link{act_data_list}} for an
output in list form. Request can be for current positions, current balances,
or current day orders. For historical orders, see \code{\link{order_search}}.
}
\examples{
\dontrun{

# Get stored refresh token
refreshToken = readRDS('/secure/location/')

# Generate a new access token
accessToken = auth_new_accessToken(refreshToken, 'consumerKey')

# Passing the accessToken is optional. The default will return balances
balances = act_data_df()
positions = act_data_df('positions',accessToken)
orders = act_data_df('orders')

}
}
