% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{act_position_list}
\alias{act_position_list}
\title{Get Position Data associated with TD Brokerage Account}
\usage{
act_position_list(accountNumber, accessToken = NULL)
}
\arguments{
\item{accountNumber}{A TD Brokerage Account associated with the Access Token}

\item{accessToken}{A valid Access Token must be set using auth_get_access. 
The most recent access token will be used by default unless one is manually
passed into the function}
}
\value{
a list of all account details with balance information
}
\description{
The output will be a list of all details related to current positions
for a TD Ameritrade account. Use act_position_df for a cleaner output
in a dataframe.
}
\examples{
\dontrun{

### A valid refresh token can be fed into the function below for a new Access Token
positions = td_act_position('accessToken','123456789')

}
}
