% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{auth_init_loginURL}
\alias{auth_init_loginURL}
\title{Auth Step 1: Initial Log In URL}
\usage{
auth_init_loginURL(callbackURL, consumerKey)
}
\arguments{
\item{callbackURL}{User generated Callback URL associated with registered TD app}

\item{consumerKey}{TD generated Consumer key associated with registered TD app}
}
\value{
login url to grant app permission to TD Brokerage account
}
\description{
Create URL for initial login to grant registered app permission to TD brokerage account
}
\details{
To generate the initial authorization and access the TD Ameritrade API an application 
needs to be registered at https://developer.tdameritrade.com/. Once an account
has been created, use My Apps to register an application. The two inputs needed are 
from the application, a Consumer Key provided by TD Ameritrade and a Callback URL 
which the user creates. The URL output should result in a landingpage to log into a 
TD Ameritrade Brokerage account.

The call back URL can be anything. The example below assumes the Callback URL is 
https://YourAppName. The Consumer Key is auto generated and can be found under 
My Apps > Keys. This function will use these two inputs to generate a URL where
the user can log in to their standard TD Ameritrade Brokerage Account and grant the
application access to the brokerage account, enabling the API. The Authorization 
Code generated at the end of the log in process will feed into auth_init_refreshToken.
For questions, please reference the guide at 
https://developer.tdameritrade.com/content/authentication-faq
}
\examples{
\dontrun{

### Visit the URL below to log in
loginURL = auth_init_loginURL('https://AppURL',
                              'consumerKey')

}
}
