% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{auth_init_refreshToken}
\alias{auth_init_refreshToken}
\title{Auth Step 2: Obtain Initial Refresh Token}
\usage{
auth_init_refreshToken(callbackURL, consumerKey, authcode_url)
}
\arguments{
\item{callbackURL}{User generated Callback URL associated with registered TD app}

\item{consumerKey}{TD generated Consumer key associated with registered TD app}

\item{authcode_url}{Authorization URL Code generated from a successful log in to the auth_init_loginURL}
}
\value{
refresh token that is valid for 90 days
}
\description{
Get an initial Refresh Token using the Authorization code URL
}
\details{
Once a URL has been generated using auth_init_loginURL, a user can visit that URL
to log into a TD brokerage account, granting access to the TD app. A successful 
log in will result in an authorization code embedded in the URL which will be fed
into this function. Once the button "Allow" is pressed, the user will be
redirected, potentially to "This site can't be reached". This indicates a
successful log in. The URL of the page contains the authorization code.
Paste the entire URL, not just the authorization code, into this function.
The code will be an extremely long alpha numeric string.
The output of this function will be a refresh token which will be used
to gain access to the TD Brokerage account(s) going forward, avoiding auth_init. 
The Refresh Token will last for 90 days, but the user can use auth_new_refreshToken 
to reset the token before expiration. If the Refresh Token expires, a new
authorization code will need to be generated by logging into the URL from 
auth_init_loginURL. 

The Refresh Token output should be saved in a very safe location, but
also accessible. It will be needed to generate an Access Token (which 
expire every 30 minutes) which is used for general account access.
}
\examples{
\dontrun{

### The URL after a successful login can be fed into the function below
refreshToken = auth_init_refreshToken('https://myTDapp','consumerKey',
                                      'https://myTDapp/?code=Auhtorizationcode')

saveRDS(refreshToken,'/secure/location/')

}
}
