% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{auth_new_accessToken}
\alias{auth_new_accessToken}
\title{Auth Step 3: Get Access Token}
\usage{
auth_new_accessToken(refreshToken, consumerKey)
}
\arguments{
\item{refreshToken}{An existing refresh token generated using auth_init_refreshToken or auth_new_refreshToken}

\item{consumerKey}{TD generated Consumer key associated with registered TD app}
}
\value{
Access Token that is valid for 30 minutes. By default it is stored in options.
}
\description{
Get a new Access Token using a valid Refresh Token
}
\details{
An Access Token is required for most functions within rameritrade. 
It serves as a user log in for a TD Brokerage account. The token is valid for 
30 minutes and allows the user to place trades, get account information,
get orider history, pull historical stock prices, etc. A Refresh Token is 
required to generate an Access Token. Functions auth_init_refreshToken or 
auth_new_refreshToken can be used to generate Refresh Tokens which stay valid
for 90 days. The Consumer Key is generated automatically when an App is 
registered on https://developer.tdameritrade.com/.  
By default, the Access Token is stored and will automatically be passed to
downstream function. However, the user can also submit an access token 
manually if multiple tokens are in use (for example: when managing more than
one log in.)
}
\examples{
\dontrun{

### A valid refresh token can be fed into the function below for a new Access Token
currefreshToken = readRDS('/secure/location/')
accessToken = auth_new_accessToken(currefreshToken,'APPCONSUMERKEY')

}
}
