% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{market_hours}
\alias{market_hours}
\title{Get Market Hours}
\usage{
market_hours(
  marketType = c("EQUITY", "OPTION", "BOND", "FUTURE", "FOREX"),
  accessToken = NULL
)
}
\arguments{
\item{marketType}{The asset class to pull:
'EQUITY','OPTION','BOND','FUTURE','FOREX'. Default is EQUITY}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
List output of times and if the current date is a trading day
}
\description{
Returns a list output for current day and specified market that details the
trading window
}
\examples{
\dontrun{

# Access Token must be set using auth_new_accessToken
# Market hours for the current date
market_hours()
market_hours('2020-06-24', 'OPTION')

}
}
