% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{option_chain}
\alias{option_chain}
\title{Get Options Chain}
\usage{
option_chain(
  ticker,
  strikes = 10,
  inclQuote = TRUE,
  startDate = Sys.Date(),
  endDate = Sys.Date() + months(12),
  accessToken = NULL
)
}
\arguments{
\item{ticker}{underlying ticker for the options chain}

\item{strikes}{the number of strikes above and below the current strike}

\item{inclQuote}{include pricing details (will be delayed if account is set for delayed quotes)}

\item{startDate}{the start date for expiration (should be greater than or equal to today). format yyyy-mm-dd}

\item{endDate}{the end date for expiration. format yyyy-mm-dd}

\item{accessToken}{A valid Access Token must be set using auth_new_accessToken. 
The most recent access token will be used by default unless one is manually
passed into the function}
}
\value{
Data frame of a tickers options chain
}
\description{
Return a Data Frame containing the options chain for the specified ticker
}
\examples{
\dontrun{

### Pull all option contracts expiring over the next 6 months 
### with 5 strikes above and below the at-the-money price
option_chain(ticker='SPY',strikes=5,endDate = Sys.Date() + months(6))

}
}
