% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{order_cancel}
\alias{order_cancel}
\title{Cancel an Open Order}
\usage{
order_cancel(orderId, accountNumber, accessToken = NULL)
}
\arguments{
\item{orderId}{A valid TD Ameritrade Order ID}

\item{accountNumber}{The account number associated with the Order ID and Access Token}

\item{accessToken}{A valid Access Token must be set using auth_new_accessToken.
The most recent access token will be used by default unless one is manually
passed into the function}
}
\value{
order URL
}
\description{
Pass an order ID and Account number to cancel an existing open order
}
\examples{
 \dontrun{

order_cancel(orderId=123456789,accountNumber=987654321)

}
}
