% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{order_detail}
\alias{order_detail}
\title{Get Details for a Single Order}
\usage{
order_detail(orderId, accountNumber, accessToken = NULL)
}
\arguments{
\item{orderId}{A valid TD Ameritrade Order ID}

\item{accountNumber}{The account number associated with the Order ID and Access Token}

\item{accessToken}{A valid Access Token must be set using auth_new_accessToken.
The most recent access token will be used by default unless one is manually 
passed into the function}
}
\value{
list of order details
}
\description{
Pass an order ID and Account number to get details such as status,
quantity, ticker, executions (if applicable), etc.
}
\examples{
\dontrun{

order_detail(orderId=123456789,accountNumber=987654321)

}
}
