% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{order_search}
\alias{order_search}
\title{Search for orders by date}
\usage{
order_search(
  accountNumber,
  startDate = Sys.Date() - months(1),
  endDate = Sys.Date(),
  maxResults = 50,
  orderStatus = "",
  accessToken = NULL
)
}
\arguments{
\item{accountNumber}{The account number associated with the Access Token}

\item{startDate}{Orders from a certain date. Will not pull back orders older than 60 days. format yyyy-mm-dd}

\item{endDate}{Filter orders that occurred before a certain date. format yyyy-mm-dd}

\item{maxResults}{the max results to return in the query}

\item{orderStatus}{search by order status (ACCEPTED, FILLED, EXPIRED, CANCELED, REJECTED, etc)}

\item{accessToken}{A valid Access Token must be set using auth_new_accessToken.
The most recent access token will be used by default unless one is manually
passed into the function}
}
\value{
a list of three objects: a jsonlite formatted extract of all orders, 
all entered orders with details, a data frame of all executed orders with the executions
}
\description{
Search for orders associated with a TD account over the previous 60 days. The result
is a list of three objects: a jsonlite formatted extract of all orders, 
all entered orders with details, a data frame of all executed orders with the executions
}
\examples{
\dontrun{

order_search(accountNumber=987654321, startDate = Sys.Date()-days(5),
             maxResult = 100, orderStatus = 'FILLED')

}
}
