% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{price_history_single}
\alias{price_history_single}
\title{Get price history for a single security}
\usage{
price_history_single(
  ticker = "AAPL",
  startDate = Sys.Date() - months(1),
  endDate = Sys.Date(),
  freq = c("daily", "1min", "5min", "10min", "15min", "30min"),
  accessToken = NULL
)
}
\arguments{
\item{ticker}{a single ticker}

\item{startDate}{the Starting point of the data}

\item{endDate}{the Ending point of the data}

\item{freq}{the frequency of the interval. Can be daily, 1min, 5min, 10min,
15min, or 30min}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
a tibble of price data
}
\description{
Open, Close, High, Low, and Volume for a specific security
}
\details{
Pulls price history for a single security based on the parameters that
include a date range and frequency of the interval. Depending on the
frequency interval, data can only be pulled back to a certain date. For
example, at a one minute interval, data can only be pulled for 30-35 days.
Prices are adjusted for splits but not dividends.
}
\examples{

\dontrun{

# Set the access token and a provide a vector of one or more tickers
refreshToken = readRDS('/secure/location/')
accessToken = auth_new_accessToken(refreshToken, 'consumerKey')
tickHist5min = price_history_single('TSLA', freq='5min')

# The default is daily. Access token is optional once it's been set
tickHistDay = price_history_single('SPY', startDate = '1990-01-01')

}
}
\seealso{
Other pricing functions: 
\code{\link{price_hisotry_mult}()},
\code{\link{price_quote_df}()},
\code{\link{price_quote_list}()}
}
\concept{pricing functions}
