% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{price_quote_df}
\alias{price_quote_df}
\title{Get Quotes for specified tickers in data frame form}
\usage{
price_quote_df(tickers = c("AAPL", "MSFT"), accessToken = NULL)
}
\arguments{
\item{tickers}{One or more tickers}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
a data frame with quote details for each valid ticker submitted
}
\description{
Enter tickers for real time or delayed quotes as data frame
}
\details{
Quotes may be delayed depending on agreement with TD Ameritrade. If the
account is set up for real-time quotes then this will return real-time.
Otherwise the quotes will be delayed.
}
\examples{
\dontrun{

# Get stored refresh token
refreshToken = readRDS('/secure/location/')

# generate a new access token
accessToken = auth_new_accessToken(refreshToken, 'consumerKey')

# Pass one or more tickers as a vector
# accessToken is optional once it is set
quoteSPY = price_quote_df('SPY')
quoteList = price_quote_df(c('GOOG','TSLA'), accessToken)

}
}
