% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rameritrade.R
\docType{package}
\name{rameritrade}
\alias{rameritrade}
\title{rameritrade: A package for using the TD Ameritrade API}
\description{
The rameritrade package provides four categories of important functions:
\enumerate{ 
    \item authenticating into a TD Brokerage account
    \item working with orders 
    \item getting account information 
    \item pulling pricing/market data
}
}
\details{
Details for the functions can be found within the function help pages. 
Detailed examples also available within the README.
}
