% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{symbol_detail}
\alias{symbol_detail}
\title{Get ticker details}
\usage{
symbol_detail(ticker, accessToken = NULL)
}
\arguments{
\item{ticker}{a valid ticker or symbol}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
data frame of ticker details
}
\description{
Get identifiers and fundamental data for a specific ticker
}
\examples{
\dontrun{

# Details for Apple
symbol_detail('AAPL')

}
}
