% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_act_position.R
\name{td_act_position}
\alias{td_act_position}
\title{Get Position Data associated with TD Brokerage Account}
\usage{
td_act_position(accessToken, accountNumber)
}
\arguments{
\item{accessToken}{A valid Access Token from td_auth_getaccess}

\item{accountNumber}{A TD Brokerage Account associated with the Access Token}
}
\value{
a list of all account details with balance information
}
\description{
The output will be a list of all details related to current positions
for a TD Ameritrade account. Use td_act_positionDF for a cleaner output
in a dataframe.
}
