% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_auth_getaccess.R
\name{td_auth_getaccess}
\alias{td_auth_getaccess}
\title{Get a new Access Token using a valid Refresh Token}
\usage{
td_auth_getaccess(refreshToken, consumerKey)
}
\arguments{
\item{refreshToken}{An existing refresh token generated using td_auth_initrefresh or td_auth_getrefresh}

\item{consumerKey}{TD generated Consumer key associated with registered TD app}
}
\value{
Access Token that is valid for 30 minutes
}
\description{
An Access Token is required for most functions within rameritrade. 
It serves as a user log in for a TD Brokerage account. The token is valid for 
30 minutes and allows the user to place trades, get account information,
get orider history, pull historical stock prices, etc. A Refresh Token is 
required to generate an Access Token. Functions td_auth_initrefresh or 
td_auth_getrefresh can be used to generate Refresh Tokens which stay valid
for 90 days. The Consumer Key is generated automatically when an App is 
registered on https://developer.tdameritrade.com/.
}
\examples{
\dontrun{

### A valid refresh token can be fed into the function below for a new Access Token
accessToken = td_auth_getaccess('CurrentRefreshToken','CONSUMERKEY')

}
}
