% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_prc_quote.R
\name{td_prc_quote}
\alias{td_prc_quote}
\title{Get Quotes for specified tickers}
\usage{
td_prc_quote(accessToken, tickers = c("AAPL", "MSFT"))
}
\arguments{
\item{accessToken}{A valid Access Token from td_auth_getaccess}

\item{tickers}{One or more tickers}
}
\value{
a list of data with quote details for each valid ticker submitted
}
\description{
Quotes may be delayed depending on agreement with TD Ameritrade. If the account
is set up for real-time quotes then this will return real-time. Otherwise the 
quotes will be delayed.
}
