% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{td_symbolDetail}
\alias{td_symbolDetail}
\title{Get ticker details}
\usage{
td_symbolDetail(ticker, accessToken = NULL)
}
\arguments{
\item{ticker}{a valid ticker or symbol}

\item{accessToken}{A valid Access Token must be set using
\code{\link{td_auth_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
data frame of ticker details
}
\description{
Get identifiers and fundamental data for a specific ticker
}
\examples{
\dontrun{

# Details for Apple
td_symbolDetail('AAPL')

}
}
