% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{transact_search}
\alias{transact_search}
\title{Search for all Transaction types}
\usage{
transact_search(
  accountNumber,
  startDate = Sys.Date() - months(1),
  endDate = Sys.Date(),
  transType = "All",
  accessToken = NULL
)
}
\arguments{
\item{accountNumber}{The TD brokerage account number associated with the
Access Token}

\item{startDate}{Transactions after a certain date. Will not pull back
transactions older than 1 year. format yyyy-mm-dd}

\item{endDate}{Filter transactions that occurred before a certain date.
format yyyy-mm-dd}

\item{transType}{Filter for a specific Transaction type. No entry will return
all types. For example: TRADE, CASH_IN_OR_CASH_OUT, CHECKING, DIVIDEND,
INTEREST, OTHER}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
a jsonlite data frame of transactions
}
\description{
Can pull trades as well as transfers, dividend reinvestment, interest, etc.
Any activity associated with the account.
}
\examples{
\dontrun{

# Access Token must be set using auth_new_accessToken
# Transactions for the last 5 days
transact_search(accountNumber = 987654321, 
                startDate = Sys.Date()-days(5))

}
}
