% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{auth_init_refreshToken}
\alias{auth_init_refreshToken}
\title{Auth Step 2: Obtain Initial Refresh Token}
\usage{
auth_init_refreshToken(callbackURL, consumerKey, authcode_url)
}
\arguments{
\item{callbackURL}{User generated Callback URL associated with registered TD
app}

\item{consumerKey}{TD generated Consumer key associated with registered TD
app. Essentially an API key.}

\item{authcode_url}{Authorization URL Code generated from a successful log in
to the \code{\link{auth_init_loginURL}}}
}
\value{
Refresh Token that is valid for 90 days
}
\description{
Get an initial Refresh Token using the Authorization Code
}
\details{
Once a URL has been generated using \code{\link{auth_init_loginURL}}, a user
can visit that URL to log into a TD brokerage account, granting the TD app
access to the account. Once the button "Allow" is pressed, the user will be
redirected, potentially to "This site can't be reached". This indicates a
successful log in. The URL of this page contains the Authorization Code.
Paste the entire URL, not just the Authorization Code, into this function.
The code will be an extremely long alpha numeric string. The output of this
function will be a Refresh Token which will be used to gain access to the TD
Brokerage account(s) going forward, avoiding the auth_init functions going
forward. The Refresh Token will last for 90 days, but the user can use
\code{\link{auth_new_refreshToken}} to reset the token before expiration. If
the Refresh Token expires, a new Authorization Code will need to be generated
by logging into the URL from \code{\link{auth_init_loginURL}}

The Refresh Token output should be saved in a very safe location, but also
accessible. It will be needed to generate an Access Token, which is used for
general account access. The Access Token expire every 30 minutes.
}
\examples{
\dontrun{

### The URL after a successful login can be fed into the function below
refreshToken = auth_init_refreshToken('https://myTDapp','consumerKey',
                                      'https://myTDapp/?code=Auhtorizationcode')

saveRDS(refreshToken,'/secure/location/')

}
}
\seealso{
\code{\link{auth_init_loginURL}} for login url,
  \code{\link{auth_init_refreshToken}} for initial Refresh Token,
  \code{\link{auth_new_accessToken}} for a new Access Token,
  \code{\link{auth_new_refreshToken}} to reset an existing Refresh Token
  before expiration

Other authentication functions: 
\code{\link{auth_init_loginURL}()},
\code{\link{auth_new_accessToken}()},
\code{\link{auth_new_refreshToken}()}
}
\concept{authentication functions}
