% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{market_hours}
\alias{market_hours}
\title{Get Market Hours}
\usage{
market_hours(
  marketDate = Sys.Date(),
  marketType = c("EQUITY", "OPTION", "BOND", "FUTURE", "FOREX"),
  accessToken = NULL
)
}
\arguments{
\item{marketDate}{The market date to pull details for}

\item{marketType}{The asset class for hour:
'EQUITY','OPTION','BOND','FUTURE','FOREX'}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
List output of times and if the specified date is a trading day
}
\description{
Returns a list output for a specified day and market that details the trading
window for that day
}
\examples{
\dontrun{

### Market hours for the current date
market_hours()

}
}
