% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{order_detail}
\alias{order_detail}
\title{Get Details for a Single Order}
\usage{
order_detail(orderId, accountNumber, accessToken = NULL)
}
\arguments{
\item{orderId}{A valid TD Ameritrade Order ID}

\item{accountNumber}{The TD brokerage account number associated with the
Access Token}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
list of order details
}
\description{
Pass an order ID and Account number to get details such as status, quantity,
ticker, executions (if applicable), account, etc.
}
\examples{
\dontrun{

# Get stored refresh token
refreshToken = readRDS('/secure/location/')

# generate a new access token
accessToken = auth_new_accessToken(refreshToken, 'consumerKey')

# Get order details for a single order
# Passing Access Token is optional once it's been set
order_detail(orderId = 123456789, accountNumber = 987654321)

}
}
