% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{price_history_single}
\alias{price_history_single}
\title{Get price history for a single security}
\usage{
price_history_single(
  ticker = "AAPL",
  startDate = Sys.Date() - months(1),
  endDate = Sys.Date(),
  freq = c("daily", "1min", "5min", "10min", "15min", "30min"),
  accessToken = NULL
)
}
\arguments{
\item{ticker}{a single ticker}

\item{startDate}{the Starting point of the data}

\item{endDate}{the Ending point of the data}

\item{freq}{the frequency of the interval. Can be daily, 1min, 5min, 10min,
15min, or 30min}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
a tibble of price data
}
\description{
Pulls price history for a single security based on the parameters that
include a date range and frequency of the interval. Depending on the
frequency interval, data can only be pulled back to a certain date. For
example, at a one minute interval, data can only be pulled for 30-35 days
}
\examples{

\dontrun{

### Use a valid access token and a vector of one or more tickers
tickHist5min = price_history_single(c('GOOG','TSLA'),freq='5min')

}
}
