% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{price_quote_df}
\alias{price_quote_df}
\title{Get Quotes for specified tickers in data frame form}
\usage{
price_quote_df(tickers = c("AAPL", "MSFT"), accessToken = NULL)
}
\arguments{
\item{tickers}{One or more tickers}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent access token will be
used by default unless one is manually passed into the function}
}
\value{
a data frame with quote details for each valid ticker submitted
}
\description{
Quotes may be delayed depending on agreement with TD Ameritrade. If the
account is set up for real-time quotes then this will return real-time.
Otherwise the quotes will be delayed.
}
\examples{
\dontrun{

### Use a valid access token and a vector of one or more tickers
quoteList = price_quote_df(c('GOOG','TSLA'),accessToken)

}
}
