% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{price_quote_list}
\alias{price_quote_list}
\title{Get Quotes for specified tickers in List form}
\usage{
price_quote_list(tickers = c("AAPL", "MSFT"), accessToken = NULL)
}
\arguments{
\item{tickers}{One or more tickers}

\item{accessToken}{A valid Access Token must be set using
\code{\link{auth_new_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
a list of data with quote details for each valid ticker submitted
}
\description{
Enter tickers for real time or delayed quotes returned as a list
}
\details{
Quotes may be delayed depending on agreement with TD Ameritrade. If the
account is set up for real-time quotes then this will return real-time.
Otherwise the quotes will be delayed.
}
\examples{
\dontrun{

# Get stored refresh token
refreshToken = readRDS('/secure/location/')

# generate a new access token
accessToken = auth_new_accessToken(refreshToken, 'consumerKey')

# Pass one or more tickers as a vector
# accessToken is optional once it is set
quoteSPY = price_quote_list('SPY')
quoteList = price_quote_list(c('GOOG','TSLA'), accessToken)

}
}
\seealso{
Other pricing functions: 
\code{\link{price_hisotry_mult}()},
\code{\link{price_history_single}()},
\code{\link{price_quote_df}()}
}
\concept{pricing functions}
